# server/build/swagger.json
{
	"components": {
...
		"schemas": {
...
			"Generation": {
				"properties": {
					"generationId": {
						"type": "number",
						"format": "double"
					},
					"description": {
						"type": "string"
					},
					"applicationStart": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"applicationEnd": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"selectionWeDateStart": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"selectionWeDateEnd": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"wwDateStart": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"wwDateEnd": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"infoEveningVisitors": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"doorCode": {
						"type": "string",
						"nullable": true
					},
					"electionStart": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"electionEnd": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"mentors": {
						"items": {
							"$ref": "#/components/schemas/Member"
						},
						"type": "array"
					},
					"members": {
						"items": {
							"$ref": "#/components/schemas/Member"
						},
						"type": "array"
					},
					"traineeApplicants": {
						"items": {
							"$ref": "#/components/schemas/TraineeApplicant"
						},
						"type": "array"
					},
					"internalProjects": {
						"items": {
							"$ref": "#/components/schemas/InternalProject"
						},
						"type": "array"
					}
				},
				"required": [
					"generationId",
					"description",
					"applicationStart",
					"applicationEnd",
					"selectionWeDateStart",
					"selectionWeDateEnd",
					"wwDateStart",
					"wwDateEnd",
					"infoEveningVisitors",
					"doorCode",
					"electionStart",
					"electionEnd",
					"mentors",
					"members",
					"traineeApplicants",
					"internalProjects"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"InternalProject": {
				"properties": {
					"internalProjectId": {
						"type": "number",
						"format": "double"
					},
					"generationId": {
						"type": "number",
						"format": "double"
					},
					"projectName": {
						"type": "string"
					},
					"abbreviation": {
						"type": "string"
					},
					"kickoff": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"offerAtEv": {
						"type": "boolean"
					},
					"zpAtEv": {
						"type": "boolean"
					},
					"zpHeld": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"apAtEv": {
						"type": "boolean"
					},
					"apHeld": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"dlAtEv": {
						"type": "boolean"
					},
					"qualityManagers": {
						"items": {
							"$ref": "#/components/schemas/Member"
						},
						"type": "array"
					},
					"members": {
						"items": {
							"$ref": "#/components/schemas/Member"
						},
						"type": "array"
					},
					"generation": {
						"$ref": "#/components/schemas/Generation"
					}
				},
				"required": [
					"internalProjectId",
					"generationId",
					"projectName",
					"abbreviation",
					"kickoff",
					"offerAtEv",
					"zpAtEv",
					"zpHeld",
					"apAtEv",
					"apHeld",
					"dlAtEv",
					"qualityManagers",
					"members",
					"generation"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"MemberStatusDto": {
				"properties": {
					"name": {
						"type": "string"
					},
					"memberStatusId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"memberStatusId"
				],
				"type": "object",
				"description": "Type of the member status dto"
			},
...
			"MentorDto": {
				"properties": {
					"lastname": {
						"type": "string"
					},
					"firstname": {
						"type": "string"
					},
					"memberId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"lastname",
					"firstname",
					"memberId"
				],
				"type": "object",
				"description": "Type of the mentor (partial member)"
			},
...
			"MembersFieldDto": {
				"properties": {
					"memberStatus": {
						"$ref": "#/components/schemas/MemberStatusDto"
					},
					"lastname": {
						"type": "string"
					},
					"firstname": {
						"type": "string"
					},
					"memberId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"lastname",
					"firstname",
					"memberId"
				],
				"type": "object",
				"description": "Type of the member field"
			},
			"InternalProjectDto": {
				"properties": {
					"qualityManagers": {
						"items": {
							"$ref": "#/components/schemas/MembersFieldDto"
						},
						"type": "array",
						"nullable": true
					},
					"members": {
						"items": {
							"$ref": "#/components/schemas/MembersFieldDto"
						},
						"type": "array",
						"nullable": true
					},
					"dlAtEv": {
						"type": "boolean"
					},
					"apHeld": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"apAtEv": {
						"type": "boolean"
					},
					"zpHeld": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"zpAtEv": {
						"type": "boolean"
					},
					"offerAtEv": {
						"type": "boolean"
					},
					"kickoff": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"abbreviation": {
						"type": "string"
					},
					"projectName": {
						"type": "string"
					},
					"generationName": {
						"type": "string"
					},
					"generation": {
						"type": "number",
						"format": "double"
					},
					"internalProjectId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"qualityManagers",
					"members",
					"dlAtEv",
					"apHeld",
					"apAtEv",
					"zpHeld",
					"zpAtEv",
					"offerAtEv",
					"kickoff",
					"abbreviation",
					"projectName",
					"generationName",
					"generation",
					"internalProjectId"
				],
				"type": "object",
				"description": "Type of the internal project of a member"
			},
			"TraineeChoiceDto": {
				"properties": {
					"choice_department3": {
						"type": "number",
						"format": "double"
					},
					"choice_department2": {
						"type": "number",
						"format": "double"
					},
					"choice_department1": {
						"type": "number",
						"format": "double"
					},
					"choice_department": {
						"type": "number",
						"format": "double"
					},
					"choice_internalProject3": {
						"type": "number",
						"format": "double"
					},
					"choice_internalProject2": {
						"type": "number",
						"format": "double"
					},
					"choice_internalProject1": {
						"type": "number",
						"format": "double"
					},
					"choice_internalProject": {
						"type": "number",
						"format": "double"
					},
					"choice_mentor3": {
						"type": "number",
						"format": "double"
					},
					"choice_mentor2": {
						"type": "number",
						"format": "double"
					},
					"choice_mentor1": {
						"type": "number",
						"format": "double"
					},
					"choice_mentor": {
						"type": "number",
						"format": "double"
					},
					"lastname": {
						"type": "string"
					},
					"firstname": {
						"type": "string"
					},
					"memberID": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"choice_department3",
					"choice_department2",
					"choice_department1",
					"choice_department",
					"choice_internalProject3",
					"choice_internalProject2",
					"choice_internalProject1",
					"choice_internalProject",
					"choice_mentor3",
					"choice_mentor2",
					"choice_mentor1",
					"choice_mentor",
					"lastname",
					"firstname",
					"memberID"
				],
				"type": "object",
				"description": "Type of the trainee choice of a member"
			},
			"TraineeMotivationDto": {
				"properties": {
					"internalProject3Motivation": {
						"type": "string"
					},
					"internalProject2Motivation": {
						"type": "string"
					},
					"internalProject1Motivation": {
						"type": "string"
					},
					"memberID": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"internalProject3Motivation",
					"internalProject2Motivation",
					"internalProject1Motivation",
					"memberID"
				],
				"type": "object",
				"description": "Type of the motivation of a trainee"
			},
			"UpdateVotingDeadlinesRequestDto": {
				"properties": {
					"electionEnd": {
						"type": "string",
						"format": "date-time"
					},
					"electionStart": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"electionEnd",
					"electionStart"
				],
				"type": "object",
				"description": "Type of the request to update the voting deadlines of a generation"
			},
			"TraineeAssignmentDto": {
				"properties": {
					"departmentID": {
						"type": "number",
						"format": "double"
					},
					"mentorID": {
						"type": "number",
						"format": "double"
					},
					"ipID": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"departmentID",
					"mentorID",
					"ipID"
				],
				"type": "object",
				"description": "Type of the assignment of a trainee to an internal project, mentor and department"
			},
			"TraineeProgressDto": {
				"properties": {
					"excelBasics": {
						"type": "boolean"
					},
					"safetyTraining": {
						"type": "boolean"
					},
					"dataPrivacyTraining": {
						"type": "boolean"
					},
					"strategyAndOrganisation": {
						"type": "boolean"
					},
					"msPowerpoint": {
						"type": "boolean"
					},
					"departmentQualityManagement": {
						"type": "boolean"
					},
					"departmentNetwork": {
						"type": "boolean"
					},
					"departmentFinanceAndLaw": {
						"type": "boolean"
					},
					"acquisitionNegotiationTechnique": {
						"type": "boolean"
					},
					"rhetoricPresentationTechnique": {
						"type": "boolean"
					},
					"projectManagement": {
						"type": "boolean"
					},
					"dlAtEv": {
						"type": "boolean"
					},
					"apHeld": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"apAtEv": {
						"type": "boolean"
					},
					"zpHeld": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"zpAtEv": {
						"type": "boolean"
					},
					"offerAtEv": {
						"type": "boolean"
					},
					"abbreviation": {
						"type": "string"
					},
					"projectName": {
						"type": "string"
					},
					"internalProjectID": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"generationID": {
						"type": "number",
						"format": "double"
					},
					"memberStatus": {
						"$ref": "#/components/schemas/MemberStatusDto"
					},
					"lastname": {
						"type": "string"
					},
					"firstname": {
						"type": "string"
					},
					"memberID": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"excelBasics",
					"safetyTraining",
					"dataPrivacyTraining",
					"strategyAndOrganisation",
					"msPowerpoint",
					"departmentQualityManagement",
					"departmentNetwork",
					"departmentFinanceAndLaw",
					"acquisitionNegotiationTechnique",
					"rhetoricPresentationTechnique",
					"projectManagement",
					"dlAtEv",
					"apHeld",
					"apAtEv",
					"zpHeld",
					"zpAtEv",
					"offerAtEv",
					"abbreviation",
					"projectName",
					"internalProjectID",
					"generationID",
					"memberStatus",
					"lastname",
					"firstname",
					"memberID"
				],
				"type": "object",
				"description": "Type of the trainee progress with the internal project milestones and mandatory workshops feedback"
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "server",
		"version": "0.1.0",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/members/{id}/status": {
			"patch": {
				"operationId": "UpdateMemberStatus",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Updates the status of an existing member\r\nUpdate can be done by members with certain permission\r\nUpdate can be done by member himself with an additional permission",
				"summary": "Update the status of a member",
				"tags": [
					"Members"
				],
				"security": [
					{
						"jwt": [
							"1"
						]
					}
				],
				"parameters": [
					{
						"description": "The id of the member to update",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "The new status of the member",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"memberStatus": {
										"type": "string"
									}
								},
								"required": [
									"memberStatus"
								],
								"type": "object",
								"description": "The new status of the member"
							},
							"example": {
								"memberStatus": "aktives Mitglied"
							}
						}
					}
				}
			}
		},
...
		"/trainees": {
			"get": {
				"operationId": "GetTrainees",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/MembersFieldDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Retrieves all current trainees",
				"summary": "Get current trainees",
				"tags": [
					"Trainees"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/trainees/ip/{id}": {
			"get": {
				"operationId": "GetIP",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InternalProjectDto"
								}
							}
						}
					}
				},
				"description": "Retrieves a single internal project by id",
				"summary": "Get an internal project",
				"tags": [
					"Trainees"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"description": "id of the internal project",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateIP",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Updates an internal project by id",
				"summary": "Update an internal project",
				"tags": [
					"Trainees"
				],
				"security": [
					{
						"jwt": [
							"15"
						]
					}
				],
				"parameters": [
					{
						"description": "id of the internal project",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "updated internal project",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/InternalProjectDto",
								"description": "updated internal project"
							},
							"example": {
								"DLBeiEV": true,
								"APGehalten": "2021-01-01",
								"APBeiEV": true,
								"ZPGehalten": "2021-01-01",
								"ZPBeiEV": true,
								"AngebotBeiEV": true,
								"kickoff": "2021-01-01",
								"kuerzel": "string",
								"projektname": "string-long",
								"generationsBezeichnung": "Wintersemester 19/20",
								"generation": 15,
								"internesProjektID": 62,
								"projektmitglieder": [
									{
										"mitgliedID": 8478,
										"vorname": "Kellan",
										"nachname": "Mclaughlin"
									},
									{
										"mitgliedID": 8748,
										"vorname": "Mason",
										"nachname": "Vinson"
									}
								],
								"qualitaetsmanager": [
									{
										"mitgliedID": 8320,
										"vorname": "Radhika",
										"nachname": "Norton"
									},
									{
										"mitgliedID": 8333,
										"vorname": "Miruna",
										"nachname": "Decker"
									}
								]
							}
						}
					}
				}
			}
		},
		"/trainees/generations/{id}/trainee-choices": {
			"get": {
				"operationId": "GetTraineeChoicesOfGeneration",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TraineeChoiceDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Retrieves choices of mentor, internal project and department of all trainees of given generation",
				"summary": "Get trainee choices",
				"tags": [
					"Trainees"
				],
				"security": [
					{
						"jwt": [
							"14"
						]
					}
				],
				"parameters": [
					{
						"description": "id of the generation",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/trainees/generations": {
			"get": {
				"operationId": "GetGenerations",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Generation"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Retrieves information about all generations",
				"summary": "Get generations",
				"tags": [
					"Trainees"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/trainees/generations/{id}/motivation": {
			"get": {
				"operationId": "GetTraineeMotivation",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TraineeMotivationDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Gets letter of motivation form trainees of given generation",
				"summary": "Get trainee motivations",
				"tags": [
					"Trainees"
				],
				"security": [
					{
						"jwt": [
							"14"
						]
					}
				],
				"parameters": [
					{
						"description": "id of the generation",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/trainees/generations/{id}/set-deadline": {
			"post": {
				"operationId": "SetElectionDeadline",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Sets \"wahl_start\" and \"wahl_ende\" for generation",
				"summary": "Set voting deadline",
				"tags": [
					"Trainees"
				],
				"security": [
					{
						"jwt": [
							"14"
						]
					}
				],
				"parameters": [
					{
						"description": "id of the generation",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "start and end date of the voting",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateVotingDeadlinesRequestDto",
								"description": "start and end date of the voting"
							},
							"example": {
								"electionStart": "2021-01-01 18:00:00",
								"electionEnd": "2021-01-31 20:00:00"
							}
						}
					}
				}
			}
		},
		"/trainees/{id}/assignment": {
			"patch": {
				"operationId": "SetTraineeAssignment",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Sets choices of internesprojekt, mentor and department of member",
				"summary": "Set trainee assignment",
				"tags": [
					"Trainees"
				],
				"security": [
					{
						"jwt": [
							"14"
						]
					}
				],
				"parameters": [
					{
						"description": "id of the trainee",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "assignment of the trainee",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TraineeAssignmentDto",
								"description": "assignment of the trainee"
							},
							"example": {
								"ipID": 62,
								"mentorID": 8167,
								"departmentID": 1
							}
						}
					}
				}
			}
		},
		"/trainees/generations/{id}/mentors": {
			"get": {
				"operationId": "GetMentorsOfGeneration",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/MentorDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Retrieves all internal projects of a generation",
				"summary": "Get internal projects",
				"tags": [
					"Trainees"
				],
				"security": [
					{
						"jwt": [
							"14"
						]
					}
				],
				"parameters": [
					{
						"description": "id of the generation",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/trainees/generations/{id}/add-mentor/{memberID}": {
			"post": {
				"operationId": "AddMentor",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Adds a mentor to a generation",
				"summary": "Add mentor",
				"tags": [
					"Trainees"
				],
				"security": [
					{
						"jwt": [
							"14"
						]
					}
				],
				"parameters": [
					{
						"description": "id of the generation",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "id of the mentor",
						"in": "path",
						"name": "memberID",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/trainees/generations/{id}/internal-projects": {
			"get": {
				"operationId": "GetInternalProjectsOfGeneration",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/InternalProjectDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Retrieves all internal projects of a generation",
				"summary": "Get internal projects",
				"tags": [
					"Trainees"
				],
				"security": [
					{
						"jwt": [
							"14"
						]
					}
				],
				"parameters": [
					{
						"description": "id of the generation",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/trainees/ips/current": {
			"get": {
				"operationId": "GetCurrentIPs",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/InternalProjectDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Retrieves all current IPs",
				"summary": "Get current IPs",
				"tags": [
					"Trainees"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/trainees/ips/all": {
			"get": {
				"operationId": "GetAllIPs",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/InternalProjectDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Retrieves all IPs",
				"summary": "Get all IPs",
				"tags": [
					"Trainees"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/trainees/generations/{id}/trainee-progress": {
			"get": {
				"operationId": "GetTraineeProgress",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/TraineeProgressDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Retrieves all milestones of internal projects and workshop feedback of a generation",
				"summary": "Get progress of the trainee generation",
				"tags": [
					"Trainees"
				],
				"security": [
					{
						"jwt": [
							"14"
						]
					}
				],
				"parameters": [
					{
						"description": "id of the generation",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/trainees/admission/{id}": {
			"post": {
				"operationId": "AdmitTrainee",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Admits a trainee which sets the active since state of the member to the current date and\r\nupdates the member status",
				"summary": "Admit a trainee",
				"tags": [
					"Trainees"
				],
				"security": [
					{
						"jwt": [
							"18"
						]
					}
				],
				"parameters": [
					{
						"description": "id of the trainee",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}